lab2-airflow-orchestration/requirements.txt
# ============================================
# MWAA 2.7.2 Compatible Requirements
# ============================================
# NOTE: Do NOT include apache-airflow - it's pre-installed!

# AWS Providers
apache-airflow-providers-amazon==8.13.0

# LLM & Embeddings
openai==1.7.2
langchain==0.1.0
langchain-openai==0.0.5

# Vector Database
pinecone-client==3.0.3

# Data & Utilities
pandas==2.1.4
numpy==1.26.3
requests==2.31.0

# Wikipedia (choose ONE - using wikipedia, not Wikipedia-API)
wikipedia==1.4.0

# Models
pydantic==2.5.3

# Database
psycopg2-binary==2.9.9
